Enemies take shake and swing damage from player
Enemies on player arms can take shake damage

Enemies on that are not will take damage on hit

Enemies fly on death

Enemies destroyed after 5 seconds

+++ Assets/Scripts/Alejandro/Enemies/SO_Base/SO_EnemySettings.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "New enemy settings", menuName = "Scriptable Objects/Enemy settings")]
 public class SO_EnemySettings : ScriptableObject
 {
-    public int MaxHealth;
+    public float MaxHealth;
     public float MoveSpeed;
 }
